+++ src/utils.h
@@ -237,7 +237,7 @@ extern SEXP strings_unique;
 extern SEXP strings_universal;
 extern SEXP strings_check_unique;
 extern SEXP strings_key;
-extern SEXP strings_pos;
+extern SEXP strings_loc;
 extern SEXP strings_val;
 extern SEXP strings_group;
 extern SEXP strings_length;