+++ tests/testthat/test-group.R
@@ -136,19 +136,19 @@ test_that("print method is useful", {
   expect_known_output(print(x), file = test_path("test-type-group-rle.txt"))
 })
 
-# group pos --------------------------------------------------------------
+# group loc --------------------------------------------------------------
 
 test_that("can locate unique groups of an empty vector", {
-  out <- vec_group_pos(integer())
+  out <- vec_group_loc(integer())
 
   expect_s3_class(out, "data.frame")
   expect_equal(out$key, integer())
-  expect_equal(out$pos, list())
+  expect_equal(out$loc, list())
 })
 
 test_that("can locate unique groups of a data frame", {
   df <- data_frame(x = c(1, 1, 1, 2, 2), y = c("a", "a", "b", "a", "b"))
-  out <- vec_group_pos(df)
+  out <- vec_group_loc(df)
 
   expect_equal(nrow(out), 4L)
   expect_equal(out$key, vec_unique(df))
@@ -156,46 +156,46 @@ test_that("can locate unique groups of a data frame", {
 
 test_that("can locate unique groups of a data frame with a list column", {
   df <- data_frame(x = list(1:2, 1:2, "a", 5.5, "a"))
-  out <- vec_group_pos(df)
+  out <- vec_group_loc(df)
 
   expect_equal(nrow(out), 3L)
   expect_equal(out$key, vec_unique(df))
 })
 
 test_that("`x` must be a vector", {
-  expect_error(vec_group_pos(environment()), class = "vctrs_error_scalar_type")
+  expect_error(vec_group_loc(environment()), class = "vctrs_error_scalar_type")
 })
 
 test_that("`key` column retains full type information", {
   x <- factor(letters[c(1, 2, 1)], levels = letters[1:3])
-  out <- vec_group_pos(x)
+  out <- vec_group_loc(x)
 
   expect_equal(levels(out$key), levels(x))
 })
 
-test_that("vec_group_pos takes the equality proxy", {
+test_that("vec_group_loc takes the equality proxy", {
   local_comparable_tuple()
   x <- tuple(c(1, 2, 1), 1:3)
-  expect_equal(vec_group_pos(x)$key, x[1:2])
-  expect_equal(vec_group_pos(x)$pos, list(c(1L, 3L), 2L))
+  expect_equal(vec_group_loc(x)$key, x[1:2])
+  expect_equal(vec_group_loc(x)$loc, list(c(1L, 3L), 2L))
 
   x <- as.POSIXlt(new_datetime(c(1, 2, 1)))
-  expect_equal(vec_group_pos(x)$key, x[1:2])
-  expect_equal(vec_group_pos(x)$pos, list(c(1L, 3L), 2L))
+  expect_equal(vec_group_loc(x)$key, x[1:2])
+  expect_equal(vec_group_loc(x)$loc, list(c(1L, 3L), 2L))
 })
 
-test_that("vec_group_pos takes the equality proxy recursively", {
+test_that("vec_group_loc takes the equality proxy recursively", {
   local_comparable_tuple()
 
   x <- tuple(c(1, 2, 1, 1), 1:4)
   df <- data_frame(x = x)
 
-  expect <- data_frame(key = vec_slice(df, c(1, 2)), pos = list(c(1L, 3L, 4L), 2L))
+  expect <- data_frame(key = vec_slice(df, c(1, 2)), loc = list(c(1L, 3L, 4L), 2L))
 
-  expect_equal(vec_group_pos(df), expect)
+  expect_equal(vec_group_loc(df), expect)
 })
 
-test_that("vec_group_pos works with different encodings", {
+test_that("vec_group_loc works with different encodings", {
   encs <- encodings()
-  expect_identical(nrow(vec_group_pos(encs)), 1L)
+  expect_identical(nrow(vec_group_loc(encs)), 1L)
 })